Refactored GameSearchListComponent
